lib socks_gssapi

+++ lib/socks_gssapi.c
@@ -283,7 +283,7 @@ CURLcode Curl_SOCKS5_gssapi_negotiate(struct Curl_cfilter *cf,
     }
 
     result = Curl_blockread_all(cf, data, (char *)gss_recv_token.value,
-                                gss_recv_token.length, &actualread);
+                                (ssize_t)gss_recv_token.length, &actualread);
 
     if(result || (actualread != us_length)) {
       failf(data, "Failed to receive GSS-API authentication token.");
@@ -472,7 +472,7 @@ CURLcode Curl_SOCKS5_gssapi_negotiate(struct Curl_cfilter *cf,
     return CURLE_OUT_OF_MEMORY;
   }
   result = Curl_blockread_all(cf, data, (char *)gss_recv_token.value,
-                              gss_recv_token.length, &actualread);
+                              (ssize_t)gss_recv_token.length, &actualread);
 
   if(result || (actualread != us_length)) {
     failf(data, "Failed to receive GSS-API encryption type.");
